fix: separated reused ast traversal methods
